medical_assistant/requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
anthropic==0.18.1
together==1.2.7
langdetect==1.0.9

